requirements.txt
absl-py==2.1.0
accelerate==0.28.0
aiohttp @ file:///home/conda/feedstock_root/build_artifacts/aiohttp_1710511536979/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1667935791922/work
annotated-types @ file:///home/conda/feedstock_root/build_artifacts/annotated-types_1696634205638/work
asttokens==2.4.1
async-timeout @ file:///home/conda/feedstock_root/build_artifacts/async-timeout_1691763562544/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1704011227531/work
av==12.0.0
bigmodelvis==0.0.1
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1695989787169/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1707022139797/work/certifi
cffi==1.16.0
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1692311806742/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.9
data==0.4
dataclasses @ file:///home/conda/feedstock_root/build_artifacts/dataclasses_1628958434797/work
datasets @ file:///home/conda/feedstock_root/build_artifacts/datasets_1709395865330/work
decorator==5.1.1
decord==0.6.0
deepspeed @ file:///home/conda/feedstock_root/build_artifacts/deepspeed_1686347801988/work
dill @ file:///home/conda/feedstock_root/build_artifacts/dill_1706434688412/work
einops==0.6.1
evaluate==0.4.2
exceptiongroup==1.2.0
executing==2.0.1
faiss==1.7.4
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1711394622191/work
fonttools==4.50.0
frozenlist @ file:///home/conda/feedstock_root/build_artifacts/frozenlist_1702645481127/work
fsspec @ file:///home/conda/feedstock_root/build_artifacts/fsspec_1707102468451/work
ftfy==6.1.1
funcsigs==1.0.2
fvcore==0.1.5.post20221221
gin==0.1.6
gin-config==0.5.0
grpcio==1.62.1
hjson @ file:///home/conda/feedstock_root/build_artifacts/hjson-py_1660536441309/work
huggingface_hub @ file:///home/conda/feedstock_root/build_artifacts/huggingface_hub_1711468615105/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1701026962277/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1710971335535/work
iopath==0.1.10
ipdb==0.13.13
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.3
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1691577114857/work
kiwisolver==1.4.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpi4py==3.1.5
mpmath==1.3.0
multidict @ file:///home/conda/feedstock_root/build_artifacts/multidict_1707040698785/work
multiprocess @ file:///home/conda/feedstock_root/build_artifacts/multiprocess_1706514640841/work
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1707225380409/work/dist/numpy-1.26.4-cp310-cp310-linux_x86_64.whl#sha256=51131fd8fc130cd168aecaf1bc0ea85f92e8ffebf211772ceb16ac2e7f10d7ca
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
open-clip-torch==2.24.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1710075952259/work
pandas==2.0.3
parameterized==0.9.0
parso==0.8.3
peft==0.3.0
pexpect==4.9.0
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1675487172403/work
plyfile==1.0.3
portalocker==2.8.2
prompt-toolkit==3.0.43
protobuf==5.26.0
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1705722392846/work
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo @ file:///home/conda/feedstock_root/build_artifacts/py-cpuinfo_1666774466606/work
pyarrow==15.0.2
pyarrow-hotfix @ file:///home/conda/feedstock_root/build_artifacts/pyarrow-hotfix_1700596371886/work
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.22
pydantic==1.10.14
pydantic_core @ file:///home/conda/feedstock_root/build_artifacts/pydantic-core_1708700740324/work
pydicom==2.4.4
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.1.2
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
pytorchvideo==0.1.5
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1706886791323/work
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1695373428874/work
regex==2022.10.31
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
rich==13.7.1
rouge-score==0.1.2
sacremoses @ file:///home/conda/feedstock_root/build_artifacts/sacremoses_1651557636210/work
safetensors @ file:///home/conda/feedstock_root/build_artifacts/safetensors_1707377218239/work
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.2.2
sentencepiece==0.2.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
soundfile==0.12.1
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.4.0
timm @ file:///home/conda/feedstock_root/build_artifacts/timm_1669264598736/work
tokenizers==0.13.3
tomli==2.0.1
torch==1.12.1
torchaudio==0.12.1
torcheval==0.0.7
torchvision==0.13.1
tqdm==4.64.1
traitlets==5.14.2
transformers==4.31.0
trimesh==4.2.0
triton==2.2.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1708904622550/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1707747584337/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1708239446578/work
wcwidth==0.2.13
Werkzeug==3.0.1
wfdb==4.1.2
xxhash @ file:///home/conda/feedstock_root/build_artifacts/python-xxhash_1696486308932/work
yacs==0.1.8
yarl @ file:///home/conda/feedstock_root/build_artifacts/yarl_1705508292061/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
